Add support for more than two columns
`maxSpreadColumns` specifies the maximum number of columns;
`evenSpreads` controls whether the number of columns can be an odd
number or not.

Note: I don't know what the "evenSpreads" option was for; but since
it's there, and it doesn't seem to be doing anything, I have
repurposed it for controlling whether or not the number of columns
must be even.